May updatee
We’ve updated the add factors spreadsheet to include the BEA data for Q12021 that was mistakenly not pulled in during the update; essentially no changes to the FIM output because most of those cells don’t affect parts of the spreadsheet that are currently pulled into the code. The few changes (affecting state health outlays, federal health outlays and state social benefits) are very minor and wouldn’t affect headline numbers in any significant capacity.
	We were double counting UI in state social benefits. However, because of a bug in the code, federal social benefits were being read in as 0. These two errors roughly balanced each other out—but both have been corrected.
	Gftfpu from Haver (what BEA calls “UI expansion”) does not include the Wages Lost Assistance Program (Trump UI in the add factors spreadsheet). However, the WLA was included in yptu (total UI). Since we defined federal_ui  as gftfpu and  state_ui = yptu (total ui) - gftfpu (federal_ui), we accidentally included the WLA in state UI, bumping state UI too high and federal UI too low. This does not affect the headline FIM.
	The state health add factor was slightly off in Q2 of 2021 because of a shifted formula in the add ons calculation sheet.
	Finally, we had forgotten to take 2020 Q4’s total corporate taxes number (instead of the average of the past three quarters that BEA spits out) since we don’t have new corporate taxes data until the third revision—we fixed this.
	Biggest change is to rebate checks.

Federal_rebate_checks and rebate_checks were two different variables; we didn’t pull the arpa rebate checks out of federal_rebate_checks (only did this for rebate_checks). However, we labeled both of these as non-arpa rebate checks, so when we summed up total contribution of rebate checks, we added in federal rebate checks as non-arpa rebate checks and added in the arpa rebate checks separately, thus double counting the arpa rebate checks. Furthermore, for federal_rebate_checks we used the mpc for rebate checks and for federal_rebate_checks_arp we used direct aid mpc, so we double counted the level and then applied different mpcs to them—bringing the contribution of rebate checks in Q1 up from about 10 to about 14.