docs: create pipeline diagram for how FIM inputs lead to FIM outputs
I've finally cracked the FIM! Well, at least the calculation steps. (There's still a lot of work to be done to understand how our data is gathered into a data frame for use in the FIM). Still, this Excel sheet now shows how the 24 variables in the forecast sheet lead to the 4 FIM output variables in very clear detail. With this done, I'm now prepared to refactor the entire calculations section.