refactor: delete sourcing of map_mpc_time_series.R file, which is no longer needed with caching
Instead of recalculating MPC matrices with each run of the FIM, we will now refer to cached MPC matrices, and only bother changing the cache if an MPC assumption has changed. Thus, the fiscal_impact_BETA.R file no longer needs to reference this script directly at this point in the refactoring. Later, I will build in observer functions that will call the function for building MPC matrices when it sees that an MPC input has been modified.