feat: add code chunks from fiscal_impact_BETA to bookdown documentation
This feature labels chunks from the fiscal_impact_BETA.R script and then uses those labels in the .Rmd files in the docs/technical_documentation directory to reprint chunks of code. It's a proof of concept that provides the scaffolding for me to write a full walk-through of the FIM script without copy-pasting code. It will change dynamically as the code changes, so long as the chunk labels are left in place.