refactor: remove variable postscript "1" used for debugging purposes
I added the "1" to the end of the variables so that I could run individual lines of code within the functions I was working on without overwriting the original input data frames. Removing it now that the functions are debugged. Perhaps in the future I could refactor these functions so they don't overwrite input arguments - might make any future debugging simpler.