refactor(create_consumption_deflator_growth): replace direct reference to create_placeholder_nas() function with placeholder_nas() argument
I accidentally left this code in from when I was refactoring this function. It worked before, but I think it's a better design choice to pass in a custom placeholder_nas data frame specified in the arguments then it is to create the data frame within the function. This forces the code to be more explicit on its assumptions.